fix: renamed PiontOutput.vue to PointOutputGraph.vue
