+++ flint.ui/src/views/flint/PointOutputGraph.vue
